+++ wolfcrypt/src/sp_x86_64.c
@@ -7555,7 +7555,7 @@ static void sp_256_ecc_recode_6_4(const sp_digit* k, ecc_recode_256* v)
     n = k[j];
     o = 0;
     for (i=0; i<43; i++) {
-        y = n;
+        y = (int8_t)n;
         if (o + 6 < 64) {
             y &= 0x3f;
             n >>= 6;
@@ -7569,12 +7569,12 @@ static void sp_256_ecc_recode_6_4(const sp_digit* k, ecc_recode_256* v)
         }
         else if (++j < 4) {
             n = k[j];
-            y |= (n << (64 - o)) & 0x3f;
+            y |= (uint8_t)((n << (64 - o)) & 0x3f);
             o -= 58;
             n >>= o;
         }
 
-        y += carry;
+        y += (uint8_t)carry;
         v[i].i = recode_index_4_6[y];
         v[i].neg = recode_neg_4_6[y];
         carry = (y >> 6) + v[i].neg;
@@ -8993,7 +8993,7 @@ static int sp_256_ecc_mulmod_stripe_4(sp_point_256* r, const sp_point_256* g,
 
         y = 0;
         for (j=0,x=42; j<6 && x<256; j++,x+=43) {
-            y |= ((k[x / 64] >> (x % 64)) & 1) << j;
+            y |= (int)(((k[x / 64] >> (x % 64)) & 1) << j);
         }
     #ifndef WC_NO_CACHE_RESISTANT
         if (ct) {
@@ -9008,7 +9008,7 @@ static int sp_256_ecc_mulmod_stripe_4(sp_point_256* r, const sp_point_256* g,
         for (i=41; i>=0; i--) {
             y = 0;
             for (j=0,x=i; j<6 && x<256; j++,x+=43) {
-                y |= ((k[x / 64] >> (x % 64)) & 1) << j;
+                y |= (int)(((k[x / 64] >> (x % 64)) & 1) << j);
             }
 
             sp_256_proj_point_dbl_4(rt, rt, t);
@@ -9423,7 +9423,7 @@ static int sp_256_ecc_mulmod_stripe_avx2_4(sp_point_256* r, const sp_point_256* 
 
         y = 0;
         for (j=0,x=42; j<6 && x<256; j++,x+=43) {
-            y |= ((k[x / 64] >> (x % 64)) & 1) << j;
+            y |= (int)(((k[x / 64] >> (x % 64)) & 1) << j);
         }
     #ifndef WC_NO_CACHE_RESISTANT
         if (ct) {
@@ -9438,7 +9438,7 @@ static int sp_256_ecc_mulmod_stripe_avx2_4(sp_point_256* r, const sp_point_256* 
         for (i=41; i>=0; i--) {
             y = 0;
             for (j=0,x=i; j<6 && x<256; j++,x+=43) {
-                y |= ((k[x / 64] >> (x % 64)) & 1) << j;
+                y |= (int)(((k[x / 64] >> (x % 64)) & 1) << j);
             }
 
             sp_256_proj_point_dbl_avx2_4(rt, rt, t);
@@ -9992,7 +9992,7 @@ static void sp_256_ecc_recode_7_4(const sp_digit* k, ecc_recode_256* v)
     n = k[j];
     o = 0;
     for (i=0; i<37; i++) {
-        y = n;
+        y = (int8_t)n;
         if (o + 7 < 64) {
             y &= 0x7f;
             n >>= 7;
@@ -10006,12 +10006,12 @@ static void sp_256_ecc_recode_7_4(const sp_digit* k, ecc_recode_256* v)
         }
         else if (++j < 4) {
             n = k[j];
-            y |= (n << (64 - o)) & 0x7f;
+            y |= (uint8_t)((n << (64 - o)) & 0x7f);
             o -= 57;
             n >>= o;
         }
 
-        y += carry;
+        y += (uint8_t)carry;
         v[i].i = recode_index_4_7[y];
         v[i].neg = recode_neg_4_7[y];
         carry = (y >> 7) + v[i].neg;
@@ -25889,7 +25889,7 @@ static void sp_384_ecc_recode_6_6(const sp_digit* k, ecc_recode_384* v)
     n = k[j];
     o = 0;
     for (i=0; i<65; i++) {
-        y = n;
+        y = (int8_t)n;
         if (o + 6 < 64) {
             y &= 0x3f;
             n >>= 6;
@@ -25903,12 +25903,12 @@ static void sp_384_ecc_recode_6_6(const sp_digit* k, ecc_recode_384* v)
         }
         else if (++j < 6) {
             n = k[j];
-            y |= (n << (64 - o)) & 0x3f;
+            y |= (uint8_t)((n << (64 - o)) & 0x3f);
             o -= 58;
             n >>= o;
         }
 
-        y += carry;
+        y += (uint8_t)carry;
         v[i].i = recode_index_6_6[y];
         v[i].neg = recode_neg_6_6[y];
         carry = (y >> 6) + v[i].neg;
@@ -27378,7 +27378,7 @@ static int sp_384_ecc_mulmod_stripe_6(sp_point_384* r, const sp_point_384* g,
 
         y = 0;
         for (j=0,x=47; j<8; j++,x+=48) {
-            y |= ((k[x / 64] >> (x % 64)) & 1) << j;
+            y |= (int)(((k[x / 64] >> (x % 64)) & 1) << j);
         }
     #ifndef WC_NO_CACHE_RESISTANT
         if (ct) {
@@ -27393,7 +27393,7 @@ static int sp_384_ecc_mulmod_stripe_6(sp_point_384* r, const sp_point_384* g,
         for (i=46; i>=0; i--) {
             y = 0;
             for (j=0,x=i; j<8; j++,x+=48) {
-                y |= ((k[x / 64] >> (x % 64)) & 1) << j;
+                y |= (int)(((k[x / 64] >> (x % 64)) & 1) << j);
             }
 
             sp_384_proj_point_dbl_6(rt, rt, t);
@@ -27806,7 +27806,7 @@ static int sp_384_ecc_mulmod_stripe_avx2_6(sp_point_384* r, const sp_point_384* 
 
         y = 0;
         for (j=0,x=47; j<8; j++,x+=48) {
-            y |= ((k[x / 64] >> (x % 64)) & 1) << j;
+            y |= (int)(((k[x / 64] >> (x % 64)) & 1) << j);
         }
     #ifndef WC_NO_CACHE_RESISTANT
         if (ct) {
@@ -27821,7 +27821,7 @@ static int sp_384_ecc_mulmod_stripe_avx2_6(sp_point_384* r, const sp_point_384* 
         for (i=46; i>=0; i--) {
             y = 0;
             for (j=0,x=i; j<8; j++,x+=48) {
-                y |= ((k[x / 64] >> (x % 64)) & 1) << j;
+                y |= (int)(((k[x / 64] >> (x % 64)) & 1) << j);
             }
 
             sp_384_proj_point_dbl_avx2_6(rt, rt, t);
